chore: reorganize `ScrollTo` components
