--- packages/core/src/ScrollTo/index.ts
@@ -1,3 +1,0 @@
-export * from "./Vertical";
-export * from "./Horizontal";
-export * from "./types";
