+++ packages/core/src/ScrollToHorizontal/ScrollToHorizontal.styles.tsx
