+++ packages/core/src/ScrollToHorizontal/ScrollToHorizontal.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it } from "vitest";
 
-import { HvProvider } from "../../providers/Provider";
+import { HvProvider } from "../providers/Provider";
 import { HvScrollToHorizontal } from "./ScrollToHorizontal";
 
 const Sample = () => (
