+++ packages/core/src/ScrollToVertical/VerticalScrollListItem/VerticalScrollListItem.styles.ts
@@ -1,7 +1,7 @@
 import { createClasses } from "@hitachivantara/uikit-react-utils";
 import { theme } from "@hitachivantara/uikit-styles";
 
-import { outlineStyles } from "../../../utils/focusUtils";
+import { outlineStyles } from "../../utils/focusUtils";
 
 const name = "HvVerticalScrollListItem";
