+++ packages/core/src/index.ts
@@ -102,7 +102,8 @@ export * from "./ProgressBar";
 export * from "./QueryBuilder";
 export * from "./Radio";
 export * from "./RadioGroup";
-export * from "./ScrollTo";
+export * from "./ScrollToHorizontal";
+export * from "./ScrollToVertical";
 export * from "./Section";
 export * from "./SimpleGrid";
 export * from "./Slider";
